docker-chatbot/requirements.txt
# requirements.txt
google-generativeai
pika
pymongo
requests
pdfplumber
python-dotenv  # Para carregar variáveis de ambiente em desenvolvimento
tenacity       # Para retentativas com backoff exponencial
python-json-logger # Para logging estruturado